Extract WorkspaceTurnToContainerDialog
