+++ opencti-platform/opencti-front/src/private/components/workspaces/WorkspaceHeader.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import * as R from 'ramda';
 import fileDownload from 'js-file-download';
 import { Field, Form, Formik } from 'formik';
-import { graphql, useLazyLoadQuery, useMutation } from 'react-relay';
+import { graphql } from 'react-relay';
 import makeStyles from '@mui/styles/makeStyles';
 import Chip from '@mui/material/Chip';
 import Typography from '@mui/material/Typography';
@@ -21,13 +21,7 @@ import { DotsHorizontalCircleOutline } from 'mdi-material-ui';
 import Button from '@mui/material/Button';
 import Tooltip from '@mui/material/Tooltip';
 import PropTypes from 'prop-types';
-import DialogTitle from '@mui/material/DialogTitle';
-import DialogContent from '@mui/material/DialogContent';
-import Autocomplete from '@mui/material/Autocomplete';
-import DialogActions from '@mui/material/DialogActions';
-import Dialog from '@mui/material/Dialog';
-import { useHistory } from 'react-router-dom';
-import StixDomainObjectCreation from '../common/stix_domain_objects/StixDomainObjectCreation';
+import WorkspaceTurnToContainerDialog from './WorkspaceTurnToContainerDialog';
 import {
   commitMutation,
   fetchQuery,
@@ -40,13 +34,6 @@ import WorkspacePopover from './WorkspacePopover';
 import ExportButtons from '../../../components/ExportButtons';
 import { useFormatter } from '../../../components/i18n';
 import WorkspaceManageAccessDialog from './WorkspaceManageAccessDialog';
-import ItemIcon from '../../../components/ItemIcon';
-import investigationToContainerAdd, { investigationToContainerMutation } from '../../../utils/ContainerUtils';
-
-const Transition = React.forwardRef((props, ref) => (
-  <Slide direction="up" ref={ref} {...props} />
-));
-Transition.displayName = 'TransitionSlide';
 
 const useStyles = makeStyles(() => ({
   title: {
@@ -170,74 +157,12 @@ const WorkspaceHeader = ({
       });
   };
 
-  const [displayTurnToReportOrCaseContainer, setDisplayTurnToReportOrCaseContainer] = useState(false);
-  const [containerCreation, setContainerCreation] = useState(false);
-  const [targetContainerId, setTargetContainerId] = useState('');
-  const [containers, setContainers] = useState([]);
-  const [actionsInputs, setActionsInputs] = useState({});
-  const [newValue, setNewValue] = useState('');
-  const [commitInvestigationToContainerAdd] = useMutation(investigationToContainerMutation);
-  const history = useHistory();
-  const handleChangeActionInputValues = (event, value) => {
-    if (event) {
-      event.stopPropagation();
-      event.preventDefault();
-    }
-    setActionsInputs({
-      ...actionsInputs || {},
-      values: Array.isArray(value) ? value : [value],
-    });
-    setTargetContainerId(value[0].value);
-  };
+  const [
+    displayTurnToReportOrCaseContainer,
+    setDisplayTurnToReportOrCaseContainer,
+  ] = useState(false);
 
   const handleOpenTurnToReportOrCaseContainer = () => setDisplayTurnToReportOrCaseContainer(true);
-  const handleCloseUpdate = () => {
-    setActionsInputs({});
-  };
-
-  const handleLaunchUpdate = () => {
-    handleCloseUpdate();
-    investigationToContainerAdd(workspace.id, targetContainerId, history, commitInvestigationToContainerAdd);
-  };
-
-  const searchContainersData = useLazyLoadQuery(
-    graphql`
-      query WorkspaceHeaderContainersQuery($search: String) {
-        containers(
-          search: $search
-          filters: [{ key: entity_type, values: ["Container"] }]
-      ) {
-        edges {
-          node {
-            id
-            entity_type
-            representative {
-              main
-            }
-          }
-        }
-        }
-      }
-    `,
-    { search: newValue ?? '' },
-  );
-  const searchContainers = (event, incomingValue) => {
-    if (!event) return;
-    setNewValue(incomingValue);
-    setActionsInputs({
-      ...actionsInputs,
-      inputValue: incomingValue ?? '',
-    });
-
-    const { containers: { edges } } = searchContainersData;
-    const elements = edges.map(({ node }) => node);
-    const containersFromElements = elements.map(({ representative, entity_type, id }) => ({
-      label: representative.main,
-      type: entity_type,
-      value: id,
-    }));
-    setContainers([...containersFromElements]);
-  };
 
   return (
     <div style={{ margin: variant === 'dashboard' ? '0 20px 0 20px' : 0 }}>
@@ -266,7 +191,9 @@ const WorkspaceHeader = ({
                 size="small"
                 style={{ width: 194, marginRight: 20 }}
               >
-                <InputLabel id="relative" variant="outlined">{t('Relative time')}</InputLabel>
+                <InputLabel id="relative" variant="outlined">
+                  {t('Relative time')}
+                </InputLabel>
                 <Select
                   labelId="relative"
                   value={relativeDate ?? ''}
@@ -395,10 +322,10 @@ const WorkspaceHeader = ({
         <Tooltip title={t('Turn into a new Report or Case')}>
           <ToggleButtonGroup size="small" color="primary" exclusive={true}>
             <ToggleButton
-                aria-label="Label"
-                onClick={handleOpenTurnToReportOrCaseContainer}
-                size="small"
-                value="turn-to-report-or-case"
+              aria-label="Label"
+              onClick={handleOpenTurnToReportOrCaseContainer}
+              size="small"
+              value="turn-to-report-or-case"
             >
               <Input color="primary" fontSize="small" />
             </ToggleButton>
@@ -492,108 +419,13 @@ const WorkspaceHeader = ({
           </Slide>
         </Security>
       </div>
-      <Dialog
-        PaperProps={{ elevation: 1 }}
-        fullWidth={true}
-        maxWidth="sm"
-        TransitionComponent={Transition}
-        open={displayTurnToReportOrCaseContainer}
-        onClose={() => setDisplayTurnToReportOrCaseContainer(false)}
-      >
-        <DialogTitle>{t('Turn to Report or Case')}</DialogTitle>
-        <DialogContent>
-          <StixDomainObjectCreation
-            inputValue={actionsInputs?.inputValue || ''}
-            open={containerCreation}
-            display={true}
-            speeddial={true}
-            stixDomainObjectTypes={[
-              'Report',
-              'Grouping',
-              'Case-Incident',
-              'Case-Rfi',
-              'Case-Rft',
-            ]}
-            handleClose={() => setContainerCreation(false)}
-            creationCallback={({ name, id, entity_type }) => {
-              const element = {
-                label: name,
-                value: id,
-                type: entity_type,
-              };
-              setContainers([...(containers ?? []), element]);
-              handleChangeActionInputValues(null, [
-                ...(actionsInputs?.values ?? []),
-                element,
-              ]);
-            }}
-          />
-          <Autocomplete
-            size="small"
-            fullWidth={true}
-            selectOnFocus={true}
-            autoHighlight={true}
-            getOptionLabel={(option) => (option.label ?? '')
-            }
-            value={actionsInputs?.values || []}
-            multiple={true}
-            renderInput={(params) => (
-              <MUITextField
-                {...params}
-                variant="standard"
-                label={t('Values')}
-                fullWidth={true}
-                onFocus={(event) => searchContainers(event)}
-                style={{ marginTop: 3 }}
-              />
-            )}
-            noOptionsText={t('No available options')}
-            options={containers}
-            onInputChange={(event) => searchContainers(event)}
-            inputValue={actionsInputs?.inputValue || ''}
-            onChange={(event, value) => handleChangeActionInputValues(event, value)}
-            renderOption={(props, option) => (
-              <li {...props}>
-                <div className={classes.icon}>
-                  <ItemIcon type={option.type} />
-                </div>
-                <div className={classes.text}>{option.label}</div>
-              </li>
-            )}
-            disableClearable
-          />
-          <IconButton
-              onClick={() => setContainerCreation(true)}
-              edge="end"
-              style={{ position: 'absolute', top: 68, right: 48 }}
-              size="large"
-          >
-            <Add />
-          </IconButton>
-        </DialogContent>
-        <DialogActions>
-          <Button
-              onClick={() => setDisplayTurnToReportOrCaseContainer(false)}
-          >
-            {t('Cancel')}
-          </Button>
-          <Button
-            color="secondary"
-            onClick={() => {
-              setDisplayTurnToReportOrCaseContainer(false);
-              setActionsInputs({
-                ...actionsInputs,
-                type: 'ADD',
-                fieldType: 'ATTRIBUTE',
-                field: 'container-object',
-              });
-              handleLaunchUpdate();
-            }}
-          >
-            {t('Add')}
-          </Button>
-        </DialogActions>
-      </Dialog>
+      <WorkspaceTurnToContainerDialog
+        workspace={workspace}
+        displayTurnToReportOrCaseContainer={displayTurnToReportOrCaseContainer}
+        setDisplayTurnToReportOrCaseContainer={
+          setDisplayTurnToReportOrCaseContainer
+        }
+      />
       <div className="clearfix" />
     </div>
   );
